login working properly

+++ main.py
@@ -28,6 +28,8 @@ def task_not_found(e):
 @app.route('/index/')
 @app.route('/')
 def index():
+    if session.get('user') is None:
+        redirect(url_for('login'))
     return render_template('index.html')
 
 @app.route('/register/', methods=['GET', 'POST'])
@@ -75,8 +77,16 @@ def login():
         # form did not validate or GET request
         return render_template("login.html", form=login_form)
 
+@app.route('/logout/')
+def logout():
+    if session.get('user'):
+        session.clear()
+    return redirect(url_for('login'))
+
 @app.route('/tasks/new/', methods=['GET', 'POST'])
 def new_task():
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     # If submit button was pressed
     if request.method == 'POST': 
         # set fields of task equal to values fetched from HTML form
@@ -105,6 +115,8 @@ def new_task():
 
 @app.route('/tasks/', methods=['GET', 'POST'])
 def get_tasks():
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     _tasks = db.session.query(Task).all()
     _tasks.sort(key=lambda task: not task.pinned)
     return render_template('tasks.html', tasks=_tasks)
@@ -113,11 +125,15 @@ def get_tasks():
 
 @app.route('/tasks/<task_id>/')
 def get_task(task_id):
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     a_task = db.session.query(Task).filter_by(id=task_id).one()
     return render_template('task.html', task=a_task)
 
 @app.route('/tasks/edit/<task_id>/', methods=['GET', 'POST'])
 def update_task(task_id):
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     if request.method == 'POST':
         title = request.form['title'].strip()
         text = request.form['taskText'].strip()
@@ -140,6 +156,8 @@ def update_task(task_id):
 
 @app.route('/tasks/delete/<task_id>/', methods=['POST'])
 def delete_task(task_id):
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     my_task = db.session.query(Task).filter_by(id=task_id).one()
     db.session.delete(my_task)
     db.session.commit()
@@ -147,6 +165,9 @@ def delete_task(task_id):
 
 @app.route('/tasks/pin/<task_id>', methods=['POST'])
 def pin_task(task_id):
+    print(session.get('user'))
+    if session.get('user') is None:
+        return redirect(url_for('login'))
     my_task = db.session.query(Task).filter_by(id=task_id).one()
     my_task.pinned = not my_task.pinned
     db.session.add(my_task)
